fix: separate long test file for SwiftLint
